# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API gerenciamento de reservas de veículos",
        "description": "Essa API tem como objetivo gerenciar reservas de veículos. Inicialmente desenvolvida como parte do processo seletivo para Happmobi.",
        "contact": {
            "email": "this.dev.gabriel@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/v1/list": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Listagem de veículos",
                "description": "Retorna uma listagem de todos os veículos",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/reservation": {
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Reservar veículo",
                "description": "Reserva um veículo",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ID": {
                                        "type": "number"
                                    }
                                }
                            },
                            "examples": {
                                "vehicle": {
                                    "value": {
                                        "ID": 2
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Vehicle"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "ID inválido ou inexistente."
                    },
                    "404": {
                        "description": "Veículo não encontrado"
                    },
                    "500": {
                        "description": "Erro interno"
                    }
                }
            }
        },
        "/v1/release": {
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Liberar veículo",
                "description": "Liberar um veículo",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ID": {
                                        "type": "number"
                                    }
                                }
                            },
                            "examples": {
                                "vehicle": {
                                    "value": {
                                        "ID": 2
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/Vehicle"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "ID inválido ou inexistente."
                    },
                    "404": {
                        "description": "Veículo não encontrado"
                    },
                    "500": {
                        "description": "Erro interno"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Vehicle": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "type": {
                        "type": "string"
                    },
                    "brand": {
                        "type": "string"
                    },
                    "model": {
                        "type": "string"
                    },
                    "available": {
                        "type": "boolean"
                    }
                }
            }
        }
    }
}